Add rectify-terminal-io.patch
During GRUB terminal input and output setup, we need to make sure that
graphics and serial console work properly for GRUB.

Using both console and serial as GRUB temrinal input and output in EFI
boot mode, routes the console to all available serial ports by default.
This also causes doubled output, since both console and serial are
writing to serial port. To fix this issue we use gfxterm and make sure
we fallback to serial output only, if gfxterm fails.

For hardware that does not have any graphics option and rely on serial
port, gfxterm would fail with error 'no suitable video mode'. Serial
console will still be enabled in this case.

If serial console is disabled, we try to load gfxterm and if it fails,
we fallback to console.

Signed-off-by: Umer Saleem <usaleem@ixsystems.com>